fix: refactored use-network-status
