+++ plugins/woocommerce-admin/client/utils/react-hooks/use-network-status.tsx
@@ -1,0 +1,28 @@
+/**
+ * External dependencies
+ */
+import { useState, useEffect } from '@wordpress/element';
+
+export const useNetworkStatus = () => {
+	const [ isNetworkOffline, setIsNetworkOffline ] = useState( false );
+
+	useEffect( () => {
+		const setOfflineBannerImage = () => {
+			setIsNetworkOffline( true );
+		};
+
+		const removeOfflineBannerImage = () => {
+			setIsNetworkOffline( false );
+		};
+
+		window.addEventListener( 'offline', setOfflineBannerImage );
+		window.addEventListener( 'online', removeOfflineBannerImage );
+
+		return () => {
+			window.removeEventListener( 'offline', setOfflineBannerImage );
+			window.removeEventListener( 'online', removeOfflineBannerImage );
+		};
+	}, [] );
+
+	return isNetworkOffline;
+};
